# [Static]: Improve sanitizePath performance